requirements.txt
discord.py
# أو python-telegram-bot
